log a more informative error when wrap-git subprojects cannot be downloaded
"ERROR: Git program not found" is both highly true, and somewhat
inscrutable. Sure, looking at the line number you can basically figure
out that subproject('something') must somehow need git to operate, but
that may not be immediately obvious.

Make mention of the fact that it is needed to "download foo.wrap".

Fixes #7764