Argument against source file list specified as wild cards does not stand.
I'm sorry, but the argument that says that dynamically looking up for 10k files is not achievable in a fraction of a second doesn't stand for a second.
First, looking up for directory changes means you only need to stat the directories, i.e. you don't need to read the directories!
Second, since you are already stating the 10k files for changes anyway, it already proves that this amount of stats is not problematic.
I've done an implementation in a python-based build system circa 2009 and it took about 0.6s for 20k files.