mark a couple of typing-only imports as noqa, to appease pyflakes
Since it cannot resolve `import typing as T` in order to figure out that
T.* is doing annotation-worthy stuff.

Since T.cast('Foo') is not actually using Foo except in an annotation
context (due to being a string) it requires extra work to resolve, and
the only thing that would currently work is actually using
'typing.cast'. However, we have decided to not use it except as T...

Since this import is only imported during mypy it's not so bad to noqa
it.