flake8: extend the ignore list, do not override
Some things are disabled by default because most people are not
interested in them. But (non-obviously) adding a list of ignores to the
flake8 configuration file will actually turn back on anything you didn't
explicitly specify.

This reduces the number of flake8 warnings by about 100, a vast chunk of
which are

E704 multiple statements on one line (def)

which turn out to actually be T.overload or typing definitions with

def func(): ...

which is understandably quite reasonable to permit.