install_man: add support for building man pages via internal custom_target
This is completely terrible, because really you should use a
custom_target and then feed it to install_man. However, that has been
decided to never happen. So, reluctantly, if install_man won't accept a
custom_target, then let install_man copy the entire interface and
internally make one.

This solves a massive usability problem where custom targets with
install_dir are strictly inferior to install_man() which automatically
handles sectioned or localized subdirectories of mandir, and results in
nontrivial code to painfully work around this. One alternative approach
was to hack in @MANDIR@ support to install_dir itself (which is itself
an inconsistent interface), but after adding a locale kwarg to
install_man it is no longer possible to infer all information from the
filename alone, so that's a total wash.

Fixes:
1550
2244
4778
4981
5935
8291