Revert "msubprojects: if fetching the remote fails, gracefully fallback to local copy"
This reverts commit 4568482316f63c48559b97c9a4771740ed61eca6.

As it turns out, the rationale for this was completely bogus. This
command doesn't re-apply the patch_directory etc. and in fact there is
no command whatsoever that does this. So, this command does not have two
purposes, and we are not making one of them more robust -- instead it
has one purpose, and we are making it lie about whether it failed.

Instead of trying to freeload off of this command, we will just add
another command to properly apply patch overlays.