backend/vs: Move Microsoft.Cpp.props to before ItemDefinitionGroup.
The main reason for this move is to make it easier to merge the copies of
project generation. But as far as I can tell, the Microsoft.Cpp.props import
also belongs before the ItemDefinitionGroup.

Originally the order seems to have been that way, but 431a9ea664 changed it in
the course of other changes.