rust: clean up inputs and outputs to bindgen
Make the handling of bindgen inputs consistent with the inputs and outputs of
source_strings_to_files.

However, for the first source argument reject anything that cannot
possibly be an header file.  This also fixes a mypy failure from the next
patch, since ExtractedObjects does not have a zero-argument get_outputs
method.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>