modules/gnome: use textwrap.dedent
Python is a whitespace significant language, changing indent level
implies that scope is changing. So when a string like

```python
def foo():
    a = '''
somthing
'''
    return a
```

It's visually misleading. Couple that with folding editors like vim
getting utterly confused by this, and it turns into a real pain. Using
textwrap.dedent allows us to get rid of that:

```python
def foo():
    a = texwrap.dedent(
        '''
        something
        ''')
    return a
```

But we still get the same result