Move {Executable,Test}Serialisation into own file.
This is a partial step on the way to make meson wrapped command execution
faster. A large fraction of the time of a meson_exe execution is spent
importing build and backends. But those are only needed for
ExecutableSerialisation.

The easiest way to deal with that seems to be to separate out the
serialization objects into their own file - they don't clearly seem to belong
in backends anyway.

This change doesn't yet affect the speed of doing so from mesonmain, because
that imports a lot on its own. But using meson_exe directly like
  time python -m mesonbuild.scripts.meson_exe --capture foo echo blarg
went from ~150ms to ~60ms on my machine.