guard a complicated mlog.warning inside a Feature check
This is only relevant on certain versions of meson, so do not print it
when meson_version is too low.

The message itself is not precisely a deprecation warning, since
ostensibly it may be an unlikely coding mistake. It is probably an
attempt to implement `copy: true`, but it might not be, hence "warning"
instead of "deprecation". So although we could switch this to a
FeatureDeprecated, that is not being done at this time.