hdf5 dependency: check for alternative config-tool names
Depending on whether hdf5 is compiled with parallel support, the
same config-tool program may be installed with a mysterious "p" in the
name. In this case, dependency lookup will totally fail, unless of
course you use the superior pkg-config interface in which case you get a
predictable name.

Work around this insanity by checking for both types of config-tool
name.

Fixes #9555