unit tests: Don't check quoting with multiple libs
pkgconf has a bug on MSYS2 due to which prefixes with spaces are not
handled correctly if the library has a Requires: on another library
and both have prefixes with spaces in them.

See: https://github.com/pkgconf/pkgconf/issues/238

So move the unit test to libanswer.pc instead of libfoo.pc till that
is fixed.