python system dependency: specify the correct library search directory
We have the guaranteed correct location in the extracted sysconfig
variables. This may or may not be in the default compiler search paths,
and at least on freebsd it is NOT. As is par the course for freebsd
problems, it is in /usr/local/lib which is explicitly not in the default
path, so we need to manually pass it.