dub: rework the way we find dub
The code for finding dub has a couple of problems.
1. It sets the value to False when dub can't be found (which is a rather
   old way of doing this)
2. It's duplicated between DubDependency and DlangModule
3. It does some odd stuff to copy between the class and instance

This fixes all of that by using a constructor to find dub at
DubDependency construction time, and they simply takes a reference to
that value in the module. And yes, this is using a constructor not an
initializer. It also makes use of the NonExistingExternalProgram type to
avoid the need to set the value to False, and finally does away with the
odd class to instance and back copying.