modules/gnome: allow building gtkdoc documentation during build
This fixes a long-standing bug where gtkdoc documentation is
built during the installation step. Directly reverting this
behavior is not possible because some users depend on that
bug. Aim to solve this in the long-run by introducing an option
which allows building during build and prints a warning if that
is not the case. A later meson release should deprecate the
option and finally remove it to force users to not depend on
the bug.