simplify destdir_join for readability
We can immediately short-circuit if there is no destdir, as we simply
return the prefix unchanged.

If there is some kind of destdir and the prefix contains a drive letter,
then no matter what we need to remove the drive letter before joining.
Technically, if the destdir is a relative path e.g. `destdir\` and
`C:\prefix`, we should still install to `destdir\prefix` without the
drive letter.

But... we also guarantee that destdir is an absolute path (or empty)
anyway. And even if we didn't, non-absolute destdir is a broken concept
for a variety of complicated reasons. So none of this matters in
practice.

One way or another, we don't need to actually check whether destdir is
an absolute path before cutting off the prefix drive letter.