gnome module: add global option to choose when to build gtkdoc HTML docs
It is fairly common for projects to want to make installation of HTML
docs optionkal. Many projects have custom options that use

if get_option('something')
    indented_gtkdoc_invocation
endif

which is messy, hard to read, and above all else, inconsistent with
different names for each project. Known options include
- gtk_doc
- gtk-doc
- doc
- docs

and possibly more. This is especially problematic for yielding options
to a superproject, which is an exercise in frustration when it turns out
each project has a different name.

Solve all these problems by resolving the option in the module itself.

This uses the new module options infrastructure, so the option name is
backwards compatible. Any existing uses of get_option will simply add an
extra layer of checks, but not break the option namespace.

FIXME: linkify ticket references

Fixes:

1811
8315

Related:
7444