azure ci: Bump windows image and VS versions
The Windows-2016 image is being deprecated and will be removed soon.
There is no image that contains Visual Studio 2017 anymore, and
installing it manually will be too flaky and slow.

We have a lot of test failures with VS 2022, so bumping to that will
take some work. Just move all VS 2017 jobs to VS 2019 for now.