rename lint workflow for less confusion
For a long time now, this has done 3 different things, only one of which
is mypy. But they are labeled:

- LintMypy / lint

  (runs pylint)

- LintMypy / custom_lint

  (runs a custom script)

- LintMypy / mypy

  (actually runs mypy)

This reduces the usable length of the label which isn't all that long to
begin with, is actively misleading, and even when you know what is going
on, it becomes harder to tell at a glance what failed. Change it to be
more unambiguous.