Fix order of paths passed to applications running with wine
The paths to the build directory must to be placed before the
system paths to avoid using possibly incorrect dll's.

This is done by a special sorting that divides the path into two groups,
inside the build directory and outside. The paths pointing into the build
directory are placed at the beginning of the list. Within a group, the
order of the paths is not changed.

Identical paths that are contained multiple times are merged into one
path beforehand.

Fixes #9783

Signed-off-by: Ralf Habacker <ralf.habacker@freenet.de>